store: split high-level download tests to store_download_test.go
